requirements.txt
paste>=3.4
pyramid-jinja2>=2.8
virtualenv>=20.0.10
waitress>=1.4.3